Refactored tests for better scalability as this project grows.
