+++ app/xenia/tests/script_test.go 
@@ -1,0 +1,2 @@
+// Package endpoint implements users tests for the API layer.
+package tests
