+++ app/xenia/tests/setup_test.go
@@ -1,0 +1,64 @@
+// Package endpoint implements users tests for the API layer.
+package tests
+
+import (
+	"testing"
+
+	"github.com/coralproject/xenia/app/xenia/routes"
+	"github.com/coralproject/xenia/pkg/query/qfix"
+	"github.com/coralproject/xenia/tstdata"
+
+	"github.com/ardanlabs/kit/cfg"
+	"github.com/ardanlabs/kit/db"
+	"github.com/ardanlabs/kit/db/mongo"
+	"github.com/ardanlabs/kit/tests"
+	"github.com/ardanlabs/kit/web/app"
+)
+
+var a *app.App
+
+func init() {
+	tests.Init("XENIA")
+
+	cfg := mongo.Config{
+		Host:     cfg.MustString("MONGO_HOST"),
+		AuthDB:   cfg.MustString("MONGO_AUTHDB"),
+		DB:       cfg.MustString("MONGO_DB"),
+		User:     cfg.MustString("MONGO_USER"),
+		Password: cfg.MustString("MONGO_PASS"),
+	}
+	tests.InitMongo(cfg)
+
+	a = routes.API().(*app.App)
+}
+
+//==============================================================================
+
+// TestMain helps to clean up the test data.
+func TestMain(m *testing.M) {
+	db := db.NewMGO()
+	defer db.CloseMGO()
+
+	tstdata.Generate(db)
+	defer tstdata.Drop()
+
+	loadQuery(db, "basic.json")
+	loadQuery(db, "basic_var.json")
+	defer qfix.Remove(db)
+
+	m.Run()
+}
+
+// loadQuery adds queries to run tests.
+func loadQuery(db *db.DB, file string) error {
+	qs1, err := qfix.Get(file)
+	if err != nil {
+		return err
+	}
+
+	if err := qfix.Add(db, qs1); err != nil {
+		return err
+	}
+
+	return nil
+}
