+++ pkg/query/exec_test.go
@@ -10,28 +10,11 @@ import (
 	"github.com/coralproject/xenia/pkg/script/sfix"
 	"github.com/coralproject/xenia/tstdata"
 
-	"github.com/ardanlabs/kit/cfg"
 	"github.com/ardanlabs/kit/db"
-	"github.com/ardanlabs/kit/db/mongo"
 	"github.com/ardanlabs/kit/tests"
 
 	"gopkg.in/mgo.v2/bson"
 )
-
-func init() {
-	tests.Init("XENIA")
-
-	cfg := mongo.Config{
-		Host:     cfg.MustString("MONGO_HOST"),
-		AuthDB:   cfg.MustString("MONGO_AUTHDB"),
-		DB:       cfg.MustString("MONGO_DB"),
-		User:     cfg.MustString("MONGO_USER"),
-		Password: cfg.MustString("MONGO_PASS"),
-	}
-	tests.InitMongo(cfg)
-}
-
-//==============================================================================
 
 // TestUmarshalMongoScript tests the ability to convert string based Mongo
 // commands into a bson map for processing.
