+++ pkg/query/query_test.go
@@ -8,26 +8,9 @@ import (
 	"github.com/coralproject/xenia/pkg/query"
 	"github.com/coralproject/xenia/pkg/query/qfix"
 
-	"github.com/ardanlabs/kit/cfg"
 	"github.com/ardanlabs/kit/db"
-	"github.com/ardanlabs/kit/db/mongo"
 	"github.com/ardanlabs/kit/tests"
 )
-
-func init() {
-	tests.Init("XENIA")
-
-	cfg := mongo.Config{
-		Host:     cfg.MustString("MONGO_HOST"),
-		AuthDB:   cfg.MustString("MONGO_AUTHDB"),
-		DB:       cfg.MustString("MONGO_DB"),
-		User:     cfg.MustString("MONGO_USER"),
-		Password: cfg.MustString("MONGO_PASS"),
-	}
-	tests.InitMongo(cfg)
-}
-
-//==============================================================================
 
 // TestUpsertCreateQuery tests if we can create a query record in the db.
 func TestUpsertCreateQuery(t *testing.T) {
